Expose the I2C, SPI, UART, and ARM modules
Public modules lead to a more organized API.
